count

+++ data-network/src/main/java/tv/anilibria/module/data/network/entity/app/release/FavoriteInfoResponse.kt
@@ -8,6 +8,6 @@ import com.squareup.moshi.JsonClass
  */
 @JsonClass(generateAdapter = true)
 data class FavoriteInfoResponse(
-    @Json(name = "rating") val rating: Int,
+    @Json(name = "rating") val rating: Long,
     @Json(name = "added") val isAdded: Boolean
 )
